Add tests for makealias commannd
The command makealias verify duplication by qresource and not in the
whole file. The command create alias for each resource in qrc file. In
case of duplication, the command skips the resource and wars the user at
the end.

Problems with generation of directories and files by TestQRCFile class has
been resolved.