Add qrc command to generate qrc file from a folder
This command takes a folder as argument where contained resources will
be record in the new qrc. The output command define where write the
newly created qrc file. All resources' path in the qrc file are
related to qrc's folder path. In case where the path defined in the
ouput option already exist, a prompt requests an overwritting
confirmation.