Create a basic class to handle config file
The ProjectConfig is a class generating a yaml file that will contain
all information about the project such as its name and location. If
config file already exist in the current directory or higher, the class
load it.

A verbose class has been designed to manage verbose messages. Three type of
messages are handled: INFO, WARNING and ERROR. Each of them has a
specific color.

The skeleton of the init command that initialize a PyQt5 project has
beed created.